Safari-611.1.10/LayoutTests/http/tests/security/cross-origin-indexeddb.html

<!DOCTYPE html>
<html>
<head>
    <script>
        var frames = 2;
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFramesAsText();
            internals.settings.setStorageBlockingPolicy('BlockThirdParty');
        }

        function decrement() {
            --frames;
            if (!frames && window.testRunner)
                internals.settings.setStorageBlockingPolicy('AllowAll');
        }
    </script>
</head>
<body>
    <p>Both frames can successfully open the database.</p>
    <iframe src="http://localhost:8000/security/resources/cross-origin-iframe-for-indexeddb.html" onload="decrement()"></iframe>
    <iframe src="http://127.0.0.1:8000/security/resources/cross-origin-iframe-for-indexeddb.html" onload="decrement()"></iframe>
</body>
</html>
