LayoutTests/http/tests/security/cross-origin-indexeddb-expected.txt

The first iframe below should return a security error, and the second should successfully open the database.

  

--------
Frame: '<!--frame1-->'
--------
window.indexedDB.deleteDatabase() threw an exception: SecurityError
window.indexedDB.open() threw an exception: SecurityError


--------
Frame: '<!--frame2-->'
--------
Successfully called window.indexedDB.deleteDatabase().
Successfully called window.indexedDB.open().
